+++ src/FairPlayCombinedSln/FairPlayCombinedDb/FairPlayTube/Tables/VideoWatchTime.sql
@@ -4,7 +4,7 @@
     [VideoInfoId] BIGINT NOT NULL, 
     [SessionId] UNIQUEIDENTIFIER NOT NULL, 
     [SessionStartDatetime] DATETIMEOFFSET NOT NULL, 
-    [WatchTime] INT NOT NULL, 
+    [WatchTime] FLOAT NOT NULL, 
     [WatchedByApplicationUserId] NVARCHAR(450) NULL, 
     CONSTRAINT [FK_VideoWatchTime_VideoInfo] FOREIGN KEY ([VideoInfoId]) REFERENCES [FairPlayTube].[VideoInfo]([VideoInfoId]), 
     CONSTRAINT [FK_VideoWatchTime_AspNetUsers] FOREIGN KEY ([WatchedByApplicationUserId]) REFERENCES [dbo].[AspNetUsers]([Id])
